+++ core/src/config/ConfigInit.cpp
@@ -85,6 +85,11 @@ InitConfig() {
         {"general.timezone", CreateStringConfig_("general.timezone", _MODIFIABLE, &config.general.timezone.value,
                                                  "UTC+8", is_timezone_valid, nullptr)},
         {"general.meta_uri", CreateStringConfig("general.meta_uri", &config.general.meta_uri.value, "sqlite://:@:/")},
+        {"general.stale_snapshots_count",
+         CreateIntegerConfig("general.stale_snapshots_count", 0, 100, &config.general.stale_snapshots_count.value, 0)},
+        {"general.stale_snapshots_duration",
+         CreateIntegerConfig("general.stale_snapshots_duration", 0, std::numeric_limits<int64_t>::max(),
+                             &config.general.stale_snapshots_duration.value, 10)},
 
         /* network */
         {"network.bind.address",
@@ -193,6 +198,7 @@ InitConfig() {
 }
 
 const char* config_file_template = R"(
+
 # Copyright (C) 2019-2020 Zilliz. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance
@@ -228,9 +234,19 @@ cluster:
 #                      | Keep 'dialect://:@:/', 'dialect' can be either 'sqlite' or |            |                 |
 #                      | 'mysql', replace other texts with real values.             |            |                 |
 #----------------------+------------------------------------------------------------+------------+-----------------+
+# stale_snapshots_count| Specify how many stale snapshots to be kept before GC. It  | Integer    |  0              |
+#                      | is ignored if deployed with cluster enabled                |            |                 |
+#----------------------+------------------------------------------------------------+------------+-----------------+
+# stale_snapshots_duration |                                                        | Integer    |  10             |
+#                      | Specify how long the stale snapshots can be GC'ed. The unit|            |                 |
+#                      | is second. It is only effective if deployed with cluster   |            |                 |
+#                      | enabled and cluster.role is rw                             |            |                 |
+#----------------------+------------------------------------------------------------+------------+-----------------+
 general:
   timezone: @general.timezone@
   meta_uri: @general.meta_uri@
+  stale_snapshots_count: @general.stale_snapshots_count@
+  stale_snapshots_duration: @general.stale_snapshots_duration@
 
 #----------------------+------------------------------------------------------------+------------+-----------------+
 # Network Config       | Description                                                | Type       | Default         |