+++ core/src/db/snapshot/ResourceGCEvent.h
@@ -16,6 +16,7 @@
 #include <set>
 #include <string>
 
+#include "config/ServerConfig.h"
 #include "db/snapshot/MetaEvent.h"
 #include "db/snapshot/Operations.h"
 #include "db/snapshot/ResourceHelper.h"
@@ -30,12 +31,18 @@ class ResourceGCEvent : public GCEvent {
     using Ptr = std::shared_ptr<ResourceGCEvent>;
 
     explicit ResourceGCEvent(typename ResourceT::Ptr res) : res_(res) {
+        auto is_cluster = config.cluster.enable();
+        auto role = config.cluster.role();
+        read_only_ = is_cluster && (role == ClusterRole::RO);
     }
 
     ~ResourceGCEvent() = default;
 
     Status
     Process(StorePtr store) override {
+        if (read_only_) {
+            return Status::OK();
+        }
         /* mark resource as 'deleted' in meta */
         auto sd_op = std::make_shared<SoftDeleteOperation<ResourceT>>(res_->GetID());
         STATUS_CHECK((*sd_op)(store));
@@ -93,6 +100,8 @@ class ResourceGCEvent : public GCEvent {
 
  private:
     typename ResourceT::Ptr res_;
+
+    bool read_only_ = false;
 };
 
 }  // namespace milvus::engine::snapshot