+++ core/src/db/snapshot/Snapshots.h
@@ -15,6 +15,7 @@
 #include <map>
 #include <memory>
 #include <mutex>
+#include <set>
 #include <shared_mutex>
 #include <string>
 #include <thread>
@@ -22,6 +23,8 @@
 #include "db/snapshot/SnapshotHolder.h"
 #include "db/snapshot/Store.h"
 #include "utils/Status.h"
+#include "utils/ThreadPool.h"
+#include "utils/TimerManager.h"
 
 namespace milvus::engine::snapshot {
 
@@ -60,9 +63,15 @@ class Snapshots {
     DropPartition(const ID_TYPE& collection_id, const ID_TYPE& partition_id, const LSN_TYPE& lsn);
 
     Status
+    NumOfSnapshot(const std::string& collection_name, int& num) const;
+
+    Status
     Reset();
 
     Status Init(StorePtr);
+
+    Status
+    RegisterTimers(TimerManager* mgr);
 
  public:
     Status
@@ -78,6 +87,11 @@ class Snapshots {
     Status
     DoDropCollection(ScopedSnapshotT& ss, const LSN_TYPE& lsn);
 
+    void
+    OnReaderTimer(const boost::system::error_code&);
+    void
+    OnWriterTimer(const boost::system::error_code&);
+
     Status
     LoadNoLock(StorePtr store, ID_TYPE collection_id, SnapshotHolderPtr& holder);
     Status
@@ -85,8 +99,12 @@ class Snapshots {
 
     mutable std::shared_timed_mutex mutex_;
     std::map<ID_TYPE, SnapshotHolderPtr> holders_;
-    std::map<std::string, ID_TYPE> name_id_map_;
-    std::vector<Snapshot::Ptr> to_release_;
+    std::set<ID_TYPE> alive_cids_;
+    std::map<std::string, std::set<ID_TYPE>> name_id_map_;
+    mutable std::shared_timed_mutex inactive_mtx_;
+    std::map<ID_TYPE, SnapshotHolderPtr> inactive_holders_;
+    std::set<ID_TYPE> invalid_ssid_;
+    StorePtr store_;
 };
 
 }  // namespace milvus::engine::snapshot